--- packages/circuit-ui/components/AvatarInput/AvatarInput.docs.mdx
@@ -1,18 +1,0 @@
-import { Status, Props, Story } from '../../../../.storybook/components';
-
-# AvatarInput
-
-<Status.Stable />
-
-The AvatarInput component allows users to upload avatar images. It can be used on its own or as part of a form.
-
-<Story id="components-avatarinput--base" />
-
-<Props />
-
-## Usage guidelines
-
-- **Do** use a short and concise alert message. Consider the maximum amount of characters in one line (50-60, including spaces) for an optimal readability.
-- **Do not** use the AvatarInput component if the avatar is read-only for the user, use the [Avatar](Components/Avatar) component instead.
-
-## (TBD)
