+++ packages/circuit-ui/components/ImageInput/ImageInput.docs.mdx
@@ -1,0 +1,18 @@
+import { Status, Props, Story } from '../../../../.storybook/components';
+
+# ImageInput
+
+<Status.Stable />
+
+The ImageInput component allows users to upload avatar images. It can be used on its own or as part of a form.
+
+<Story id="components-ImageInput--base" />
+
+<Props />
+
+## Usage guidelines
+
+- **Do** use a short and concise alert message. Consider the maximum amount of characters in one line (50-60, including spaces) for an optimal readability.
+- **Do not** use the ImageInput component if the avatar is read-only for the user, use the [Avatar](Components/Avatar) component instead.
+
+## (TBD)
